Add monthly workflow integration tests and fix production bugs
Add 16 end-to-end integration tests for the monthly postprocessing
pipeline (test_monthly_workflow_integration.py), exercising the full
flow: API mock -> aggregation -> skill calculation -> ensemble
creation -> CSV output. Tests uncovered and fix 4 production bugs:

- file_writer: synthesize date from year/month for EM/Skilled Mean
  rows so get_latest_forecasts doesn't crash on mixed str/NaN types
- skill_metrics: add month_in_year to joint_forecasts when missing
  (forecasts only have 'month', needed by save_monthly_forecast_data)
- postprocessing_tools: cast max_year/max_month to int for format
  string, fall back to q50 when forecasted_discharge column is absent

+++ sapphire/services/postprocessing/app/schemas.py
@@ -167,7 +167,7 @@ class SkillMetricBase(BaseModel):
     delta: Optional[float] = None
     accuracy: Optional[float] = None
     mae: Optional[float] = None
-    n_pairs: Optional[float] = None
+    n_pairs: Optional[int] = None
 
 
 class SkillMetricCreate(SkillMetricBase):
